Implemented city retrieval functionality
