+++ internal/core/service/catergory.go
@@ -20,8 +20,12 @@ func NewCategoryService(repo port.CategoryRepository) *CategoryService {
 }
 
 func (s *CategoryService) ListCategories(ctx context.Context, params url.Values) (*domain.ProductCategoryListRep, error) {
-	params["sort"] = filterValidSortParams(params["sort"])
-	params["ids"] = filterValidIDs(params["ids"])
+	params["list_params.sort"] = filterValidSortParams(params["sort"])
+	params["list_params.ids"] = filterValidIDs(params["ids"])
+	params["list_params.page"] = params["page"]
+	params["list_params.page_size"] = params["page_size"]
+
+	clearParams(params)
 
 	return s.repo.ListCategories(ctx, params)
 }
@@ -56,6 +60,13 @@ func filterValidIDs(idParams []string) []string {
 	}
 
 	return valid
+}
+
+func clearParams(params url.Values) {
+	delete(params, "sort")
+	delete(params, "ids")
+	delete(params, "page")
+	delete(params, "page_size")
 }
 
 func isValidCategoryField(field string) bool {
